Catch "message undefined" type errors
`parseMessage` sometimes returns undefined and the only way it can
return undefined is if `messageDef` is falsey. That means a message
definition could not be found. It looks like all message definitions are
there according to the kegboard protocol so we should throw an error
for undefined message definitions by returning the ID for the message
kegboard emitted.